new file for mood.js working
